Use namespace packages for new checkpoint libraries
- Both `langgraph` and `langgraph.checkpoint` are namespace packages, so neither can contain a top-level init file
- Each new checkpointer implementation should be packaged as `langgraph.checkpoint.xyz`
