+++ libs/checkpoint/langgraph/checkpoint/sqlite/__init__.py
@@ -1,5 +1,3 @@
-import json
-import pickle
 import sqlite3
 import threading
 from contextlib import AbstractContextManager, contextmanager
@@ -8,52 +6,23 @@ from types import TracebackType
 from typing import Any, AsyncIterator, Dict, Iterator, Optional, Sequence, Tuple
 
 from langchain_core.runnables import RunnableConfig
-from langgraph_checkpoint.base import (
+from typing_extensions import Self
+
+from langgraph.checkpoint.base import (
     BaseCheckpointSaver,
     Checkpoint,
     CheckpointMetadata,
     CheckpointTuple,
+    EmptyChannelError,
     SerializerProtocol,
 )
-from langgraph_checkpoint.serde.jsonplus import JsonPlusSerializer
-from typing_extensions import Self
-
-from langgraph.channels.base import BaseChannel
-from langgraph.errors import EmptyChannelError
-
-
-class JsonPlusSerializerCompat(JsonPlusSerializer):
-    """A serializer that supports loading pickled checkpoints for backwards compatibility.
-
-    This serializer extends the JsonPlusSerializer and adds support for loading pickled
-    checkpoints. If the input data starts with b"\x80" and ends with b".", it is treated
-    as a pickled checkpoint and loaded using pickle.loads(). Otherwise, the default
-    JsonPlusSerializer behavior is used.
-
-    Examples:
-        >>> import pickle
-        >>> from langgraph.checkpoint.sqlite import JsonPlusSerializerCompat
-        >>>
-        >>> serializer = JsonPlusSerializerCompat()
-        >>> pickled_data = pickle.dumps({"key": "value"})
-        >>> loaded_data = serializer.loads(pickled_data)
-        >>> print(loaded_data)  # Output: {"key": "value"}
-        >>>
-        >>> json_data = '{"key": "value"}'.encode("utf-8")
-        >>> loaded_data = serializer.loads(json_data)
-        >>> print(loaded_data)  # Output: {"key": "value"}
-    """
-
-    def loads(self, data: bytes) -> Any:
-        if data.startswith(b"\x80") and data.endswith(b"."):
-            return pickle.loads(data)
-        return super().loads(data)
-
+from langgraph.checkpoint.serde.types import ChannelProtocol
+from langgraph.checkpoint.sqlite.utils import JsonPlusSerializerCompat, search_where
 
 _AIO_ERROR_MSG = (
     "The SqliteSaver does not support async methods. "
     "Consider using AsyncSqliteSaver instead.\n"
-    "from langgraph.checkpoint.aiosqlite import AsyncSqliteSaver\n"
+    "from langgraph.checkpoint.sqlite.aio import AsyncSqliteSaver\n"
     "Note: AsyncSqliteSaver requires the aiosqlite package to use.\n"
     "Install with:\n`pip install aiosqlite`\n"
     "See https://langchain-ai.github.io/langgraph/reference/checkpoints/asyncsqlitesaver"
@@ -476,7 +445,7 @@ class SqliteSaver(BaseCheckpointSaver, AbstractContextManager):
         """
         raise NotImplementedError(_AIO_ERROR_MSG)
 
-    def get_next_version(self, current: Optional[str], channel: BaseChannel) -> str:
+    def get_next_version(self, current: Optional[str], channel: ChannelProtocol) -> str:
         """Generate the next version ID for a channel.
 
         This method creates a new version identifier for a channel based on its current version.
@@ -498,82 +467,3 @@ class SqliteSaver(BaseCheckpointSaver, AbstractContextManager):
         except EmptyChannelError:
             next_h = ""
         return f"{next_v:032}.{next_h}"
-
-
-def _metadata_predicate(
-    metadata_filter: Dict[str, Any],
-) -> Tuple[Sequence[str], Sequence[Any]]:
-    """Return WHERE clause predicates for (a)search() given metadata filter.
-
-    This method returns a tuple of a string and a tuple of values. The string
-    is the parametered WHERE clause predicate (excluding the WHERE keyword):
-    "column1 = ? AND column2 IS ?". The tuple of values contains the values
-    for each of the corresponding parameters.
-    """
-
-    def _where_value(query_value: Any) -> Tuple[str, Any]:
-        """Return tuple of operator and value for WHERE clause predicate."""
-        if query_value is None:
-            return ("IS ?", None)
-        elif (
-            isinstance(query_value, str)
-            or isinstance(query_value, int)
-            or isinstance(query_value, float)
-        ):
-            return ("= ?", query_value)
-        elif isinstance(query_value, bool):
-            return ("= ?", 1 if query_value else 0)
-        elif isinstance(query_value, dict) or isinstance(query_value, list):
-            # query value for JSON object cannot have trailing space after separators (, :)
-            # SQLite json_extract() returns JSON string without whitespace
-            return ("= ?", json.dumps(query_value, separators=(",", ":")))
-        else:
-            return ("= ?", str(query_value))
-
-    predicates = []
-    param_values = []
-
-    # process metadata query
-    for query_key, query_value in metadata_filter.items():
-        operator, param_value = _where_value(query_value)
-        predicates.append(
-            f"json_extract(CAST(metadata AS TEXT), '$.{query_key}') {operator}"
-        )
-        param_values.append(param_value)
-
-    return (predicates, param_values)
-
-
-def search_where(
-    config: Optional[RunnableConfig],
-    filter: Optional[Dict[str, Any]],
-    before: Optional[RunnableConfig] = None,
-) -> Tuple[str, Sequence[Any]]:
-    """Return WHERE clause predicates for (a)search() given metadata filter
-    and `before` config.
-
-    This method returns a tuple of a string and a tuple of values. The string
-    is the parametered WHERE clause predicate (including the WHERE keyword):
-    "WHERE column1 = ? AND column2 IS ?". The tuple of values contains the
-    values for each of the corresponding parameters.
-    """
-    wheres = []
-    param_values = []
-
-    # construct predicate for config filter
-    if config is not None:
-        wheres.append("thread_id = ?")
-        param_values.append(config["configurable"]["thread_id"])
-
-    # construct predicate for metadata filter
-    if filter:
-        metadata_predicates, metadata_values = _metadata_predicate(filter)
-        wheres.extend(metadata_predicates)
-        param_values.extend(metadata_values)
-
-    # construct predicate for `before`
-    if before is not None:
-        wheres.append("thread_ts < ?")
-        param_values.append(before["configurable"]["thread_ts"])
-
-    return ("WHERE " + " AND ".join(wheres) if wheres else "", param_values)
