+++ libs/checkpoint/langgraph/checkpoint/sqlite/aio.py
@@ -10,21 +10,28 @@ from typing import (
     Optional,
     Sequence,
     Tuple,
+    TypeAlias,
     TypeVar,
 )
 
-import aiosqlite
 from langchain_core.runnables import RunnableConfig
-from langgraph_checkpoint.base import (
+from typing_extensions import Self
+
+from langgraph.checkpoint.base import (
     BaseCheckpointSaver,
     Checkpoint,
     CheckpointMetadata,
     CheckpointTuple,
     SerializerProtocol,
 )
-from typing_extensions import Self
+from langgraph.checkpoint.sqlite.utils import JsonPlusSerializerCompat, search_where
 
-from langgraph.checkpoint.sqlite import JsonPlusSerializerCompat, search_where
+try:
+    import aiosqlite
+
+    AsyncConnection: TypeAlias = aiosqlite.Connection
+except ImportError:
+    AsyncConnection: TypeAlias = Any
 
 T = TypeVar("T", bound=callable)
 
@@ -84,9 +91,8 @@ class AsyncSqliteSaver(BaseCheckpointSaver, AbstractAsyncContextManager):
 
         ```pycon
         >>> import asyncio
-        >>> import aiosqlite
         >>>
-        >>> from langgraph.checkpoint.aiosqlite import AsyncSqliteSaver
+        >>> from langgraph.checkpoint.sqlite.aio import AsyncSqliteSaver
         >>> from langgraph.graph import StateGraph
         >>>
         >>> builder = StateGraph(int)
@@ -104,7 +110,7 @@ class AsyncSqliteSaver(BaseCheckpointSaver, AbstractAsyncContextManager):
         ```pycon
         >>> import asyncio
         >>> import aiosqlite
-        >>> from langgraph.checkpoint.aiosqlite import AsyncSqliteSaver
+        >>> from langgraph.checkpoint.sqlite.aio import AsyncSqliteSaver
         >>>
         >>> async def main():
         >>>     async with aiosqlite.connect("checkpoints.db") as conn:
@@ -120,13 +126,12 @@ class AsyncSqliteSaver(BaseCheckpointSaver, AbstractAsyncContextManager):
 
     serde = JsonPlusSerializerCompat()
 
-    conn: aiosqlite.Connection
     lock: asyncio.Lock
     is_setup: bool
 
     def __init__(
         self,
-        conn: aiosqlite.Connection,
+        conn: AsyncConnection,
         *,
         serde: Optional[SerializerProtocol] = None,
     ):
@@ -145,6 +150,8 @@ class AsyncSqliteSaver(BaseCheckpointSaver, AbstractAsyncContextManager):
         Returns:
             AsyncSqliteSaver: A new AsyncSqliteSaver instance.
         """
+        import aiosqlite
+
         return AsyncSqliteSaver(conn=aiosqlite.connect(conn_string))
 
     async def __aenter__(self) -> Self:
