+++ libs/checkpoint/pyproject.toml
@@ -6,7 +6,7 @@ authors = []
 license = "MIT"
 readme = "README.md"
 repository = "https://www.github.com/langchain-ai/langgraph"
-packages = [{ include = "langgraph_checkpoint" }]
+packages = [{ include = "langgraph" }]
 
 [tool.poetry.dependencies]
 python = "^3.9.0,<4.0"
@@ -20,6 +20,8 @@ pytest-asyncio = "^0.21.1"
 pytest-mock = "^3.11.1"
 pytest-watch = "^4.2.0"
 mypy = "^1.10.0"
+dataclasses-json = "^0.6.7"
+aiosqlite = "^0.20.0"
 
 [tool.pytest.ini_options]
 # --strict-markers will raise errors on unknown marks.
