+++ libs/checkpoint/tests/test_sqlite.py
@@ -1,14 +1,14 @@
 import pytest
 from langchain_core.runnables import RunnableConfig
-from langgraph_checkpoint.base import Checkpoint, CheckpointMetadata, empty_checkpoint
 
-from langgraph.channels.manager import create_checkpoint
-from langgraph.checkpoint.sqlite import (
-    _AIO_ERROR_MSG,
-    SqliteSaver,
-    _metadata_predicate,
-    search_where,
+from langgraph.checkpoint.base import (
+    Checkpoint,
+    CheckpointMetadata,
+    create_checkpoint,
+    empty_checkpoint,
 )
+from langgraph.checkpoint.sqlite import SqliteSaver
+from langgraph.checkpoint.sqlite.utils import _metadata_predicate, search_where
 
 
 class TestSqliteSaver:
@@ -116,10 +116,10 @@ class TestSqliteSaver:
 
     async def test_informative_async_errors(self):
         # call method / assertions
-        with pytest.raises(NotImplementedError, match=_AIO_ERROR_MSG):
+        with pytest.raises(NotImplementedError, match="AsyncSqliteSaver"):
             await self.sqlite_saver.aget(self.config_1)
-        with pytest.raises(NotImplementedError, match=_AIO_ERROR_MSG):
+        with pytest.raises(NotImplementedError, match="AsyncSqliteSaver"):
             await self.sqlite_saver.aget_tuple(self.config_1)
-        with pytest.raises(NotImplementedError, match=_AIO_ERROR_MSG):
+        with pytest.raises(NotImplementedError, match="AsyncSqliteSaver"):
             async for _ in self.sqlite_saver.alist(self.config_1):
                 pass
