--- libs/langgraph/langgraph/__init__.py
@@ -1,3 +1,0 @@
-from langgraph.version import __version__
-
-__all__ = ["__version__"]
