+++ libs/langgraph/langgraph/errors.py
@@ -1,3 +1,6 @@
+from langgraph.checkpoint.base import EmptyChannelError
+
+
 class GraphRecursionError(RecursionError):
     """Raised when the graph has exhausted the maximum number of steps.
 
@@ -13,13 +16,6 @@ class GraphRecursionError(RecursionError):
             {"recursion_limit": 1000},
         )
     """
-
-    pass
-
-
-class EmptyChannelError(Exception):
-    """Raised when attempting to get the value of a channel that hasn't been updated
-    for the first time yet."""
 
     pass
 
@@ -40,3 +36,12 @@ class EmptyInputError(Exception):
     """Raised when graph receives an empty input."""
 
     pass
+
+
+__all__ = [
+    "GraphRecursionError",
+    "InvalidUpdateError",
+    "GraphInterrupt",
+    "EmptyInputError",
+    "EmptyChannelError",
+]
