+++ libs/langgraph/langgraph/pregel/loop.py
@@ -22,22 +22,22 @@ from typing import (
 
 from langchain_core.callbacks import AsyncParentRunManager, ParentRunManager
 from langchain_core.runnables import RunnableConfig
-from langgraph_checkpoint.base import (
-    BaseCheckpointSaver,
-    Checkpoint,
-    CheckpointMetadata,
-    CheckpointTuple,
-    PendingWrite,
-    copy_checkpoint,
-    empty_checkpoint,
-)
 from typing_extensions import Self
 
 from langgraph.channels.base import BaseChannel
 from langgraph.channels.manager import (
     AsyncChannelsManager,
     ChannelsManager,
+)
+from langgraph.checkpoint.base import (
+    BaseCheckpointSaver,
+    Checkpoint,
+    CheckpointMetadata,
+    CheckpointTuple,
+    PendingWrite,
+    copy_checkpoint,
     create_checkpoint,
+    empty_checkpoint,
 )
 from langgraph.constants import CONFIG_KEY_READ, CONFIG_KEY_RESUMING, INPUT, INTERRUPT
 from langgraph.errors import EmptyInputError, GraphInterrupt
