+++ libs/langgraph/pyproject.toml
@@ -22,7 +22,6 @@ syrupy = "^4.0.2"
 httpx = "^0.26.0"
 pytest-watcher = "^0.4.1"
 langchain = ">=0.1.0"
-aiosqlite = "^0.19.0"
 grandalf = "^0.8"
 mypy = "^1.6.0"
 ruff = "^0.1.4"
