+++ libs/langgraph/tests/test_interruption.py
@@ -1,7 +1,6 @@
 from typing import TypedDict
 
-from langgraph_checkpoint.memory import MemorySaver
-
+from langgraph.checkpoint.memory import MemorySaver
 from langgraph.graph import END, START, StateGraph
